fix: Fix dashboard screens
This commit fixes the dashboard view. It also
seperates the registration and onboarding logic
into different files.
