+++ config/styles/components/modal.css
@@ -1,14 +1,23 @@
 .modal {
     @apply
     absolute top-0 left-0 bg-black bg-opacity-50 z-20
-    flex items-center justify-center
-    w-full h-full !mt-0
-    transition-transform ease-in-out
+    hidden items-center justify-center
+    w-full h-full max-w-full max-h-screen overflow-auto !mt-0
+    transition-transform ease-in-out p-8
     /* slide right */
     translate-x-full
 }
 
+.modal-content {
+    @apply relative w-fit max-w-lg
+    bg-white shadow-lg rounded-lg
+    p-4 md:p-6
+    transform scale-95
+    transition-all duration-300 ease-in-out
+    target:scale-100
+}
+
 .modal.active {
     @apply
-    translate-x-0
+    translate-x-0 flex
 }
