+++ config/urls.py
@@ -15,8 +15,6 @@ urlpatterns = [
     path("documents/", include(wagtaildocs_urls)),
     path("search/", search_views.search, name="search"),
     # path('', include('home.urls')),
-    path('<slug:custom_link>', restaurant_views.RestaurantDetailView.as_view()),
-
     path("api/", include("core.urls")),
     path("accounts/", include("allauth.urls")),
     path("", include(api)),
@@ -25,6 +23,11 @@ urlpatterns = [
     path("customers/", include("customers.urls")),
     path("inventory/", include("inventory.urls")),
     path("orders/", include("orders.urls")),
+    path(
+        "<slug:custom_link>",
+        restaurant_views.RestaurantDetailView.as_view(),
+        name="restaurant_detail",
+    ),
 ]
