+++ restaurants/static/js/add-employee-modal.js
@@ -1,10 +1,10 @@
 (() => {
     /**@type {HTMLDivElement} */
-    const addEmployeeModal = document.getElementById('add-employee-modal');
+    const modal = document.getElementById('add-employee-modal');
     /**@type {HTMLButtonElement} */
-    const cancelBtn = addEmployeeModal.querySelector('button.cancel');
+    const cancelBtn = modal.querySelector('button.cancel');
     /**@type {HTMLButtonElement} */
-    const addBtn = addEmployeeModal.querySelector('button.add');
+    const addBtn = modal.querySelector('button.add');
 
     /**@type {HTMLDivElement} */
     const emailInput = document.getElementById("email-input")
@@ -12,10 +12,10 @@
     const roleInput = document.getElementById("role-input")
 
     cancelBtn?.addEventListener('click', () => {
-        addEmployeeModal.classList.remove('active')
+        modal.classList.remove('active')
     })
 
     addBtn?.addEventListener('click', () => {
-        addEmployeeModal.classList.remove('active')
+        modal.classList.remove('active')
     });
 })();
