+++ restaurants/templates/restaurants/pages/settings.html
@@ -15,7 +15,6 @@
 
 {% block extra_js %}
   {{ block.super }}
-  <script src="{% static 'js/dashboard.js' %}" defer></script>
   <script src="{% static 'js/deactivate-account-modal.js' %}" defer></script>
   <script src="{% static 'js/change-password-modal.js' %}" defer></script>
 {% endblock %}
