+++ restaurants/views.py
@@ -1,62 +1,14 @@
 import os
 
-from django.conf import settings
-from django.core.files.storage import FileSystemStorage
-from django.http import HttpResponse
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.urls import reverse_lazy
 from django.views import generic
-from django.contrib.auth.mixins import LoginRequiredMixin
-from formtools.wizard.views import SessionWizardView
 
 from . import forms, models
 
 
-class SignupView(generic.CreateView):
-    form_class = forms.SignupForm
-    success_url = reverse_lazy("login")
-    template_name = "registration/restaurant/signup.html"
-
-
 class EmailVerificationView(generic.TemplateView):
     template_name = "registration/restaurant/email_verification.html"
-
-
-def show_cac_field_condition(wizard: SessionWizardView):
-    """Condition to show the CAC field"""
-    cleaned_data = wizard.get_cleaned_data_for_step("step1") or {}
-    return cleaned_data.get("business_registration_status") == "registered"
-
-
-class OnboardingWizardView(SessionWizardView):
-    form_list = [
-        ("step1", forms.OnboardingForm1),
-        ("step2A", forms.OnboardingForm2A),
-        ("step2B", forms.OnboardingForm2B),
-    ]
-    templates = {
-        "step1": "registration/restaurant/onboarding_step1.html",
-        "step2A": "registration/restaurant/onboarding_step2A.html",
-        "step2B": "registration/restaurant/onboarding_step2B.html",
-    }
-    file_storage = FileSystemStorage(location=os.path.join(settings.MEDIA_ROOT, "temp"))
-    condition_dict = {
-        "step2A": show_cac_field_condition,
-        "step2B": lambda wizard: not show_cac_field_condition(wizard),
-    }
-
-    def get_template_names(self) -> list[str]:
-        return [self.templates[self.steps.current]]
-
-    def get(self, request, *args, **kwargs):
-        return self.render(self.get_form())
-
-    def done(self, form_list, **kwargs):
-        # Save the form data to the database
-        return HttpResponse("Onboarding completed successfully")
-
-
-class OnboardingView(generic.TemplateView):
-    template_name = "registration/restaurant/onboarding.html"
 
 
 class IngredientListView(generic.ListView):
@@ -104,12 +56,12 @@ class MenuCategoryUpdateView(LoginRequiredMixin, generic.UpdateView):
     model = models.MenuCategory
     form_class = forms.MenuCategoryForm
     pk_url_kwarg = "pk"
-    
+
 
 class MenuCategoryDeleteView(LoginRequiredMixin, generic.DeleteView):
     model = models.MenuCategory
     success_url = reverse_lazy("restaurant_MenuCategory_list")
-    
+
 
 class MenuListView(LoginRequiredMixin, generic.ListView):
     model = models.Menu
@@ -130,51 +82,52 @@ class MenuUpdateView(LoginRequiredMixin, generic.UpdateView):
     model = models.Menu
     form_class = forms.MenuForm
     pk_url_kwarg = "pk"
-    
+
 
 class MenuDeleteView(LoginRequiredMixin, generic.DeleteView):
     model = models.Menu
     success_url = reverse_lazy("restaurant_Menu_list")
-    
+
 
 class AddOnListView(LoginRequiredMixin, generic.ListView):
     model = models.AddOn
     form_class = forms.ViewAddOnForm
-    
-    
+
+
 class AddOnCreateView(LoginRequiredMixin, generic.CreateView):
     model = models.AddOn
     form_class = forms.AddOnForm
-    
+
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         kwargs["user"] = self.request.user
         return kwargs
-    
+
 
 class AddOnDetailView(LoginRequiredMixin, generic.DetailView):
     model = models.AddOn
     form_class = forms.ViewAddOnForm
-    
-    
+
+
 class AddOnUpdateView(LoginRequiredMixin, generic.UpdateView):
     model = models.AddOn
     form_class = forms.AddOnForm
-    
+
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         kwargs["user"] = self.request.user
         return kwargs
-    
-    
+
+
 class AddOnDeleteView(LoginRequiredMixin, generic.DeleteView):
     model = models.AddOn
     success_url = reverse_lazy("MenuItem")
 
+
 class MenuItemListView(LoginRequiredMixin, generic.ListView):
     model = models.MenuItem
     form_class = forms.OwnerViewMenuItemForm
-    
+
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         kwargs["user"] = self.request.user
@@ -184,14 +137,14 @@ class MenuItemListView(LoginRequiredMixin, generic.ListView):
 class MenuItemCreateView(LoginRequiredMixin, generic.CreateView):
     model = models.MenuItem
     form_class = forms.MenuItemForm
-    template_name = 'menuitem_create.html'
-    #success_url = 
-    
+    template_name = "menuitem_create.html"
+    # success_url =
+
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         kwargs["user"] = self.request.user
         return kwargs
-    
+
 
 class MenuItemDetailView(LoginRequiredMixin, generic.DetailView):
     model = models.MenuItem
@@ -202,25 +155,26 @@ class MenuItemUpdateView(LoginRequiredMixin, generic.UpdateView):
     model = models.MenuItem
     form_class = forms.MenuItemForm
     pk_url_kwarg = "pk"
-    template_name = 'menuitem_update.html'
-    #success_url = 
-    
+    template_name = "menuitem_update.html"
+    # success_url =
+
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         kwargs["user"] = self.request.user
         return kwargs
-    
+
 
 class MenuItemDeleteView(LoginRequiredMixin, generic.DeleteView):
     model = models.MenuItem
     success_url = reverse_lazy("restaurant_MenuItem_list")
-    template_name = 'menuitem_delete.html'
-    #success_url = 
-    
+    template_name = "menuitem_delete.html"
+    # success_url =
+
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         kwargs["user"] = self.request.user
         return kwargs
+
 
 class RestaurantListView(generic.ListView):
     model = models.Restaurant
@@ -230,7 +184,7 @@ class RestaurantListView(generic.ListView):
 class RestaurantCreateView(LoginRequiredMixin, generic.CreateView):
     model = models.Restaurant
     form_class = forms.RestaurantForm
-    
+
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         kwargs["user"] = self.request.user
@@ -240,13 +194,13 @@ class RestaurantCreateView(LoginRequiredMixin, generic.CreateView):
 class RestaurantDetailView(generic.DetailView):
     model = models.Restaurant
     form_class = forms.RestaurantForm
-    
+
 
 class RestaurantUpdateView(LoginRequiredMixin, generic.UpdateView):
     model = models.Restaurant
     form_class = forms.RestaurantForm
     pk_url_kwarg = "pk"
-    
+
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         kwargs["user"] = self.request.user
