+++ tailwind.config.js
@@ -15,7 +15,10 @@ module.exports = {
                 'primary': '#020164',
                 'secondary': '#FE0C00',
                 'ash': 'rgba(30, 30, 30, 1)'
-            }
+            },
+            fontFamily: {
+                'poppins': ['Poppins'],
+            },
         }
     }
 }
